Fix for #136 (#137)
The type data passed to retrieve the background-operation-manager-registered command was incorrect, so commands were not returned correctly. As the commands were not returned correctly, clicking the button on the dashboard didn't wire up the click event for the command.

Resolved issue and added unit test to ensure that these sorts of commands are returned.